src/mlApp/src/requirements.txt
bson
paho-mqtt
azure-storage-blob 
pandas
matplotlib
xgboost
scikit-learn
dataclasses